Reorganizing code directories for Server-Client Architecture.
