+++ RoboCat/Chapter3.vcxproj
@@ -384,38 +384,33 @@
     </PostBuildEvent>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClInclude Include="Inc\ByteSwap.h" />
-    <ClInclude Include="Inc\Game.h" />
-    <ClInclude Include="Inc\GameObject.h" />
-    <ClInclude Include="Inc\LinkingContext.h" />
-    <ClInclude Include="Inc\MemoryBitStream.h" />
-    <ClInclude Include="Inc\OutputWindow.h" />
-    <ClInclude Include="Inc\RoboMath.h" />
-    <ClInclude Include="Inc\SocketAddress.h" />
-    <ClInclude Include="Inc\SocketAddressFactory.h" />
-    <ClInclude Include="Inc\SocketUtil.h" />
-    <ClInclude Include="Inc\StringUtils.h" />
-    <ClInclude Include="Inc\TCPSocket.h" />
-    <ClInclude Include="Inc\RoboCatPCH.h" />
-    <ClInclude Include="Inc\RoboCatShared.h" />
-    <ClInclude Include="Inc\UDPSocket.h" />
-    <ClCompile Include="Src\Game.cpp" />
-    <ClCompile Include="Src\GameObject.cpp" />
+    <ClCompile Include="Src\Util\MemoryBitStream.cpp" />
+    <ClCompile Include="Src\Util\OutputWindow.cpp" />
+    <ClCompile Include="Src\Util\RoboCatPCH.cpp" />
+    <ClCompile Include="Src\Util\SocketAddress.cpp" />
+    <ClCompile Include="Src\Util\SocketAddressFactory.cpp" />
+    <ClCompile Include="Src\Util\SocketUtil.cpp" />
+    <ClCompile Include="Src\Util\StringUtils.cpp" />
+    <ClCompile Include="Src\Util\TCPSocket.cpp" />
+    <ClCompile Include="Src\Util\UDPSocket.cpp" />
+    <ClInclude Include="Inc\Client\Game.h" />
+    <ClInclude Include="Inc\Client\GameObject.h" />
+    <ClCompile Include="Src\Client\Game.cpp" />
+    <ClCompile Include="Src\Client\GameObject.cpp" />
     <ClCompile Include="Src\Main.cpp" />
-    <ClCompile Include="Src\MemoryBitStream.cpp" />
-    <ClCompile Include="Src\OutputWindow.cpp" />
-    <ClCompile Include="Src\SocketAddress.cpp" />
-    <ClCompile Include="Src\SocketAddressFactory.cpp" />
-    <ClCompile Include="Src\SocketUtil.cpp" />
-    <ClCompile Include="Src\StringUtils.cpp" />
-    <ClCompile Include="Src\TCPSocket.cpp" />
-  </ItemGroup>
-  <ItemGroup>
-    <ClCompile Include="Src\RoboCatPCH.cpp">
-      <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Create</PrecompiledHeader>
-      <PrecompiledHeaderFile Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">RoboCatPCH.h</PrecompiledHeaderFile>
-    </ClCompile>
-    <ClCompile Include="Src\UDPSocket.cpp" />
+    <ClInclude Include="Inc\Util\ByteSwap.h" />
+    <ClInclude Include="Inc\Util\LinkingContext.h" />
+    <ClInclude Include="Inc\Util\MemoryBitStream.h" />
+    <ClInclude Include="Inc\Util\OutputWindow.h" />
+    <ClInclude Include="Inc\Util\RoboCatPCH.h" />
+    <ClInclude Include="Inc\Util\RoboCatShared.h" />
+    <ClInclude Include="Inc\Util\RoboMath.h" />
+    <ClInclude Include="Inc\Util\SocketAddress.h" />
+    <ClInclude Include="Inc\Util\SocketAddressFactory.h" />
+    <ClInclude Include="Inc\Util\SocketUtil.h" />
+    <ClInclude Include="Inc\Util\StringUtils.h" />
+    <ClInclude Include="Inc\Util\TCPSocket.h" />
+    <ClInclude Include="Inc\Util\UDPSocket.h" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
